+++ src/prefect/client/client.py
@@ -1517,25 +1517,6 @@ class Client:
         }
         self.graphql(mutation, raise_on_error=True)
 
-    def update_task_run_heartbeat(self, task_run_id: str) -> None:
-        """
-        Convenience method for heartbeating a task run.
-
-        Does NOT raise an error if the update fails.
-
-        Args:
-            - task_run_id (str): the task run ID to heartbeat
-
-        """
-        mutation = {
-            "mutation": {
-                with_args(
-                    "update_task_run_heartbeat", {"input": {"task_run_id": task_run_id}}
-                ): {"success"}
-            }
-        }
-        self.graphql(mutation, raise_on_error=True)
-
     def set_flow_run_name(self, flow_run_id: str, name: str) -> bool:
         """
         Set the name of a flow run.
