+++ src/prefect/utilities/executors.py
@@ -2,6 +2,7 @@ import cloudpickle
 import contextlib
 import copy
 import itertools
+import logging
 import multiprocessing
 import os
 import signal
@@ -11,8 +12,15 @@ import threading
 import warnings
 from concurrent.futures import ThreadPoolExecutor
 from concurrent.futures import TimeoutError as FutureTimeout
+from contextlib import contextmanager
 from functools import wraps
 from logging import Logger
+
+import prefect
+from prefect import config
+from prefect.client import Client
+from prefect.exceptions import TaskTimeoutSignal
+from prefect.utilities.logging import get_logger
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -24,11 +32,6 @@ from typing import (
     Mapping,
     Iterator,
 )
-
-import prefect
-from prefect.exceptions import TaskTimeoutSignal
-from prefect.utilities.logging import get_logger
-from prefect.utilities.threaded_heartbeat import HeartbeatThread
 
 if TYPE_CHECKING:
     import prefect.engine.runner
@@ -58,7 +61,7 @@ def run_with_heartbeat(
             use_heartbeat = self._heartbeat()
         except Exception:
             use_heartbeat = False
-            logger = get_logger()
+            logger = self.logger
             logger.exception(
                 "Heartbeat process is misconfigured.  This could result in a zombie run.",
                 exc_info=True,
@@ -138,6 +141,62 @@ def subprocess_heartbeat(heartbeat_cmd: List[str], logger: Logger) -> Iterator[N
                 logger.error(msg)
             p.kill()
             p.wait()
+
+
+class HeartbeatThread(threading.Thread):
+    def __init__(
+        self: "HeartbeatThread",
+        stop_event: "threading.Event",
+        flow_run_id: str,
+        num: int = None,
+    ) -> None:
+        threading.Thread.__init__(self)
+        # 'daemonizes' the thread, so it will terminate when all non-daemonized threads have finished
+        self.daemon = True
+        self.flow_run_id = flow_run_id
+        self.num = num
+        self.stop_event = stop_event
+
+    def run(self) -> None:
+        logger = get_logger("threaded_heartbeat")
+        client = Client()
+        iter_count = 0
+        with prefect.context(
+            {"flow_run_id": self.flow_run_id, "running_with_backend": True}
+        ):
+            with log_heartbeat_failure(logger):
+                while iter_count < (self.num or 1) and (
+                    self.stop_event.is_set() is False
+                ):
+                    send_heartbeat(self.flow_run_id, client, logger)
+                    iter_count += 1 if self.num else 0
+                    self.stop_event.wait(timeout=config.cloud.heartbeat_interval)
+
+
+def send_heartbeat(
+    flow_run_id: str, client: "prefect.client.Client", logger: "logging.Logger"
+) -> None:
+    try:  # Ignore (but log) client exceptions
+        client.update_flow_run_heartbeat(flow_run_id)
+    except Exception as exc:
+        logger.error(
+            f"Failed to send heartbeat with exception: {exc!r}",
+            exc_info=True,
+        )
+
+
+@contextmanager
+def log_heartbeat_failure(
+    logger: "logging.Logger",
+) -> Iterator[None]:
+    try:
+        yield
+    except BaseException as exc:
+        logger.error(
+            f"Heartbeat process encountered terminal exception: {exc!r}",
+            exc_info=True,
+        )
+        raise
 
 
 def run_with_thread_timeout(
