+++ tests/utilities/test_executors.py
@@ -1,12 +1,11 @@
-import os
+import cloudpickle
 import multiprocessing
+import os
+import pytest
 import sys
 import threading
 import time
 from unittest.mock import MagicMock
-
-import cloudpickle
-import pytest
 
 import prefect
 from prefect.exceptions import TaskTimeoutSignal
@@ -16,8 +15,8 @@ from prefect.utilities.executors import (
     multiprocessing_safe_run_and_retrieve,
     tail_recursive,
     RecursiveCall,
+    HeartbeatThread,
 )
-
 
 # We will test the low-level timeout handlers here and `run_task_with_timeout`
 # is covered in `tests.core.test_flow.test_timeout_actually_stops_execution`
@@ -344,3 +343,84 @@ def test_recursion_raises_when_not_decorated():
         assert a_func()
 
     assert call_checkpoints == [("a", 0), ("b", 1), ("a", 3), ("b", 4), ("a", 6)]
+
+
+def test_events_can_stop_the_heartbeat(monkeypatch):
+    Client = MagicMock()
+    monkeypatch.setattr("prefect.utilities.executors.Client", Client)
+    stop_event = threading.Event()
+    heartbeat = HeartbeatThread(stop_event, "no-flow-run-id")
+    heartbeat.start()
+    assert heartbeat.is_alive()
+    stop_event.set()
+    heartbeat.join()
+    assert heartbeat.is_alive() is False
+
+
+def test_multiple_heartbeats_can_be_independently_stopped(monkeypatch):
+    Client = MagicMock()
+    monkeypatch.setattr("prefect.utilities.executors.Client", Client)
+
+    def heartbeat_factory():
+        stop_event = threading.Event()
+        heartbeat = HeartbeatThread(stop_event, "no-flow-run-id")
+        heartbeat.start()
+        return heartbeat, stop_event
+
+    first_heartbeat, first_event = heartbeat_factory()
+    other_heartbeat, other_event = heartbeat_factory()
+    assert first_heartbeat.is_alive()
+    assert other_heartbeat.is_alive()
+    first_event.set()
+    first_heartbeat.join()
+    assert first_heartbeat.is_alive() is False
+    assert other_heartbeat.is_alive() is True
+    other_event.set()
+    other_heartbeat.join()
+    assert other_heartbeat.is_alive() is False
+
+
+def test_heartbeat_is_daemonic_by_default(monkeypatch):
+    Client = MagicMock()
+    monkeypatch.setattr("prefect.utilities.executors.Client", Client)
+    stop_event = threading.Event()
+    heartbeat = HeartbeatThread(stop_event, "no-flow-run-id")
+    assert heartbeat.isDaemon()
+
+
+def test_heartbeat_sends_signals_to_client(monkeypatch):
+    Client = MagicMock()
+    monkeypatch.setattr("prefect.utilities.executors.Client", Client)
+    stop_event = threading.Event()
+    heartbeat = HeartbeatThread(stop_event, "no-flow-run-id")
+    heartbeat.start()
+    assert heartbeat.is_alive()
+    time.sleep(0.1)
+    stop_event.set()
+    heartbeat.join()
+    assert Client().update_flow_run_heartbeat.call_count == 1
+
+
+def test_heartbeat_exceptions_are_logged_to_cloud(monkeypatch):
+    Client = MagicMock()
+    LOG_MANAGER = MagicMock()
+    monkeypatch.setattr("prefect.utilities.executors.Client", Client)
+    monkeypatch.setattr("prefect.utilities.logging.LOG_MANAGER", LOG_MANAGER)
+    Client().update_flow_run_heartbeat.side_effect = ValueError("Foo")
+
+    stop_event = threading.Event()
+    heartbeat = HeartbeatThread(stop_event, "my-special-flow-run-id")
+    heartbeat.start()
+    time.sleep(0.1)
+    stop_event.set()
+    heartbeat.join()
+
+    log = LOG_MANAGER.enqueue.call_args[0][0]
+    assert log["flow_run_id"] == "my-special-flow-run-id"
+    assert log["name"] == "prefect.threaded_heartbeat"
+    assert log["level"] == "ERROR"
+    assert "Traceback" in log["message"]
+    assert (
+        f"Failed to send heartbeat with exception: {ValueError('Foo')!r}"
+        in log["message"]
+    )
